requirements.txt
fastapi==0.116.1
uvicorn==0.33.0
feedparser==6.0.11
requests==2.32.4
PyYAML==6.0.2
python-dateutil==2.9.0.post0
